requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.2
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
numpy<2
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
pypdf
cryptography
sqlparse
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
pydantic-settings==2.1.0
sse_starlette==1.6.5
pinecone-client==3.2.2
sentence-transformers==2.7.0
Pillow
pytesseract
PyMuPDF
torch>2.2
boto3==1.34.51
aiobotocore==2.12.0
celery==5.3.6
redis==5.0.1